Move files under scanner package.
