+++ tools/blocksconvert/scanner/bigtable_index_reader.go
@@ -1,4 +1,4 @@
-package blocksconvert
+package scanner
 
 import (
 	"bytes"
